Rename send/receive options in the UI and rename associated fields
Clarify certain menu items to conform to expectations of existing wallet
users.
